fix: Fixed the interrupt flag and JUnitCore instantiation issues in ThreadsTest
In ThreadsTest, the list of interrupt flags that are no longer in use is removed, and instead, the interrupt status is recorded using a string. Meanwhile, instantiate JUnitCore as a local variable to ensure that a new instance is used for each test. The expected result of the test method was modified to verify the thread interruption status in the form of a string. The implementation of Request has been updated to handle initialization errors using anonymous inner classes.